Treewide: mark some IRQs, Kthread and worker threads as perf critical
kernel: Add API to mark IRQs and kthreads as performance critical

On devices with a CPU that contains heterogeneous cores (e.g., big.LITTLE),
it can be beneficial to place some performance-critical IRQs and kthreads
onto the performance CPU cluster in order to improve performance.

This commit adds the following APIs:
-kthread_run_perf_critical() to create and start a perf-critical kthread
-IRQF_PERF_CRITICAL to schedule an IRQ and any threads it may have onto
 performance CPUs
-PF_PERF_CRITICAL to mark a process (mainly kthread) as performance
 critical (this is used by kthread_run_perf_critical())

In order to accomodate this new API, the following changes are made:
-Performance-critical IRQs have their affinities reaffined upon exit from
 suspend (since the affinities are broken when non-boot CPUs are disabled)
-Performance-critical IRQs and their threads have their affinities reset
 upon entering suspend, so that upon immediate suspend exit (when only the
 boot CPU is online), interrupts can be processed and interrupt threads
 can be scheduled onto an online CPU (otherwise we'd hit a kernel BUG)
-Global cpu_lp_mask and cpu_perf_mask symbols are added to provide
 a cpumask of low-power and performance CPUs, respectively. This is
 set to the big.LITTLE layout used on msm8998 by default.
-do_set_cpus_allowed() and __set_cpus_allowed_ptr() are modified to enforce
 a performance-critical kthread's affinity
-Mark perf-critical IRQs with IRQD_AFFINITY_MANAGED so userspace can't mess
 with their affinity

Change-Id: I814982bd3898072dc66b04376142c5dc45b4a98f
Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>
Signed-off-by: Manish4586 <manish.n.manish45@gmail.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>

msm: kgsl: Mark IRQ and worker thread as performance critical

These items are in the critical path for rendering frames to the display,
so mark them as performance-critical.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>

msm: mdss: Mark IRQ and important kthreads as performance critical

These items are in the critical path for rendering frames to the display
(they exhibit increased CPU usage when heavy frame rendering takes place),
so mark them as performance-critical.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>

input: fingerprint: Mark IRQ as performance critical

In order to improve fingerprint wake performance, mark the fingerprint
reader's IRQ as performance-critical.

Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>