HACK: sched/tune: clamp boosts with a value of 10
To improve battery life, avoid keeping top-app boosted all the time. By
default, the Pixel 3's userspace daemons permanently boost it by 10 and
boost it to 50 on user interaction. We don't need it to stay boosted
during idle, so return it to 0 when userspace sets 10 or less.

From the commit description of Dynamic SchedTune Boost:
Dynamic SchedTune Boost occurs on-the-fly during interactions, such
as touch input. This idea was conceived because of an existing battery
drain issue whereby setting a permanent schedtune boost would cause
boosting to occur even when the phone was idling. For example, if a user
sets the /dev/stune/top-app/schedtune.boost to 10, the device may not
idle at the lowest frequency step, which can lead to higher idle and
active drain rates. Thus it should consume less power to boost only
during interactions.

Signed-off-by: kdrag0n <dragon@khronodragon.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>