arm64/configs/wayne_defconfig: Disable MSM EEPROM
Dmesg shows the following logs:

[    1.031213] msm_eeprom_platform_probe failed 2106
[    1.031942] msm_eeprom_platform_probe failed 2106
[    1.032655] msm_eeprom_platform_probe failed 2106

As it doesn't initialize, let's just disable it.

P.S: Not sure about lavender, tulip and whyred as I don't have the device physically or didn't compile the kernel for those devices yet. If your device is one of them and you are getting the same logs, you can use this commit.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>